Updating hexadecimal_format.c file

+++ hexadecimal_format.c
@@ -13,7 +13,7 @@
 */
 int print_lower_hex_format(va_list ptr, char *buffer, int *buffer_index)
 {
-	unsigned long i = va_arg(ptr, unsigned long);
+	unsigned int i = va_arg(ptr, unsigned int);
 	char *hex_format = "0123456789abcdef";
 
 	return (print_hex_format(hex_format, i, buffer, buffer_index));
@@ -29,7 +29,7 @@ int print_lower_hex_format(va_list ptr, char *buffer, int *buffer_index)
 */
 int print_upper_hex_format(va_list ptr, char *buffer, int *buffer_index)
 {
-	unsigned long int i = va_arg(ptr, unsigned long);
+	unsigned int i = va_arg(ptr, unsigned int);
 	char *hex_format = "0123456789ABCDEF";
 
 	return (print_hex_format(hex_format, i, buffer, buffer_index));
@@ -45,10 +45,10 @@ int print_upper_hex_format(va_list ptr, char *buffer, int *buffer_index)
 * Return: The number of printed characters
 */
 
-int print_hex_format(char *hex, unsigned long i, char *buff, int *buffer_ind)
+int print_hex_format(char *hex, unsigned int i, char *buff, int *buffer_ind)
 {
-	unsigned long tmp = i;
-	unsigned long divider = 1;
+	unsigned int tmp = i;
+	unsigned int divider = 1;
 	int printed_characters = 0;
 	int index;
