+++ main.h
@@ -26,7 +26,7 @@ int print_lower_hex_format(va_list ptr, char *buffer, int *buffer_index);
 int print_upper_hex_format(va_list ptr, char *buffer, int *buffer_index);
 int add_to_buffer(char *buffer, int *buffer_index, char *src, int size);
 int free_buffer(char *buf, int *buf_index);
-int print_hex_format(char *hex, unsigned long i, char *buffer, int *buffer_index);
+int print_hex_format(char *hex, unsigned int i, char *buffer, int *buffer_index);
 int print_cus_string_format(va_list ptr, char *buffer, int *buffer_index);
 int print_add_format(va_list ptr, char *buffer, int *buffer_index);
 int _len(char *c);
